+++ docs/versionhistory.rst
@@ -9,6 +9,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
   backends
 - Fixed ``KeyError`` on asyncio and curio where entering and exiting a cancel scope happens in
   different tasks
+- Fixed deprecation warnings on Python 3.8 about the ``loop`` argument of ``asyncio.Event()``
 - Dropped support for trio v0.11
 
 **1.1.0**